feat: make AI prompts configurable with editor UI
- Extracted hardcoded AI prompts to `src/config/defaultPrompts.json`.
- Implemented `PromptManager` utility for prompt loading, saving, and interpolation.
- Added Electron IPC handlers (`load-prompts`, `save-prompts`, `reset-prompts`) for persistence.
- Created `PromptEditor` component with preview and reset functionality.
- Integrated `PromptEditor` into `AISettings` dialog as a new tab.
- Refactored `aiGenerator.ts` to use dynamic prompts.
- Created `src/types/prompts.ts` for type definitions.
- Created `src/utils/promptMigration.ts` for future migration logic.
