+++ src/types/prompts.ts
@@ -1,0 +1,8 @@
+
+export interface PromptTemplate {
+  name: string;
+  description: string;
+  template: string;
+}
+
+export type PromptCollection = Record<string, PromptTemplate>;
